+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Add support for MDX files([@davemooreuws](https://github.com/davemooreuws)).
+
 ## [4.8.1] - 2021-10-31
 
 Happy Halloween!